[http] move actual tests to test_suite.cxx
Let reuse same tests with different engine
